+++ Makefile
@@ -17,6 +17,7 @@ clean:
 fclean: clean
 		docker volume ls -q | grep camagru_database | xargs -r docker volume rm
 
+re:		down fclean up
 
 logs:
 		docker compose logs
@@ -25,3 +26,6 @@ ls:
 		docker compose images
 		docker compose ps
 		docker volume ls
+
+e_enginx:
+		docker exec -it nginx_camagru sh
