+++ frontend/Public/Pages/homePage.js
@@ -1,0 +1,9 @@
+
+export default function homePage()
+{
+  return`
+      <h1 class="text-2xl font-semibold text-gray-800">Welcome to the Home Page</h1>
+      <p class="text-gray-600 mt-4">This is the home page content.</p>
+      <div id="userNameTag"></div>
+    `;
+}
