+++ frontend/Public/index.html
@@ -2,19 +2,28 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Camagru</title>
+  <!-- Tailwind CSS CDN -->
+  <script src="https://cdn.tailwindcss.com"></script>
 </head>
 
-<body>
-  <nav>
-    <a href="/" class="route" data-link>Home</a>
-    <a href="/sign_in" class="route" data-link>Sign In</a>
-  </nav>
-  <main>
-    <div id="content">
+<body class="bg-gray-50 font-sans">
+
+  <!-- Navigation bar -->
+  <nav class="bg-gray-800 text-white p-4 sticky top-0">
+    <div class="max-w-7xl mx-auto flex justify-between items-center">
+      <a href="/" class="text-xl font-bold" data-link>Camagru</a>
+      <ul class="flex space-x-4">
+        <li><a href="/" class="route text-white px-4 py-2 bg-gray-700 hover:bg-gray-600 rounded-lg" data-link>Home</a></li>
+      </ul>
     </div>
+  </nav>
+
+  <!-- Main content -->
+  <main>
+    <div id="content"></div>
   </main>
-  <div id="app"></div>
-  <script src="app.js"></script>
+  <script type="module" src="/changeContent.js"></script>
 </body>
 </html>
