+++ frontend/nginx.conf
@@ -1,3 +1,4 @@
+
 events {}
 
 http {
@@ -8,15 +9,20 @@ http {
 
 
     server_name localhost albaud.42.ch www.albaud.42.ch diskworld;
+    root /var/www/html;
 
     location / {
         # resolver 127.0.0.1 valid=30s; # Docker DNS
 
-    	root /app/Public;
     	index index.html;
         try_files $uri $uri/ /index.html;
     }
 
+    # Ensure JavaScript files are served with the correct MIME type
+    location ~* \.js$ {
+        types { application/javascript js; }
+        default_type application/javascript;
+    }
     # location /api {
     #     proxy_pass http://backend:3000/;
     # }
