--- frontend/script
@@ -1,3 +1,0 @@
-
-cp ./nginx.conf /etc/nginx/nginx.conf
-nginx -g "daemon off;"
